Generate a YAML factfile in a report-handler.
This replaces the opscodeohai_facts plugin for MCollective.
MCollective's response times are more predictable as a result,
although fact data may be stale (as it is only updated when Chef
runs).

The default configuration will not export all of the facts reported
by Ohai.  A Chef attribute defines the list of fact groups to include,
and the default leaves out some of the more verbose facts which are
less useful as fact sources.  Fact groups which were omitted from the
list are noted as one fact each: "blocked.<factgroup> = true".