Support for the native RabbitMQ connector
The development version of MCollective (2.3.x) only supports 2
connectors: ActiveMQ and RabbitMQ. The Stomp connector has been removed
so there is a need to support native RabbitMQ connector. This patch
generates a rabbitmq.cfg file when connector = 'rabbitmq'.

Along with the removal of the stomp connector, the development version
MCollective has dropped support in the configuration for the topicprefix
parameter.  This patch ensures that the parameter is declared only when
connector = 'stomp'.