Restart mcollective when the chef agent is updated.
The server doesn't automatically reload agents, so this is necessary
before changes are visible.  Note that if the chef client applying
this change is running under mcollective, the request will be
terminated (but the chef run will finish).

This approach is used to share a single pair of resource definition
between client and server recipes, which happens in the 'common'
recipe.  The service resource hasn't been defined when that recipe is
compiled, so we need to do it here.

It kinda smells - need a better pattern for this.